use Decimal package for arbitrary precision
It is dangerous to bet money on floats due to limited precision and
rounding error. We thus use the Decimal package to represent quantities
returned by the API.

When performing computation, it may be wise to use satoshis and pips
to perform calculations instead of working with floating point
numbers.